Main changes for Release v4.5.0:
* fix(mount): Set default cacheExpirationTime to 0 (#180)
* feat(mount): Add readcachemaxsize option (#145)
* feat(master): Add empty reserved files option (#176)
* fix(mount): Fix random case option parsing (#135)
* feat(chunkserver): Add IDiskManager::getDiskForGC (#169)
* test: Improve `stop_ganesha()` function (#168)
* fix(ci): Modify timeout from env (#165)
* ci: Fix tests behavior (#172)
* fix(uraft): Fix missing floating ip address
* fix(mount): Fix create time (Windows) (#167)
* feat(admin): Add list-disk-groups command (#155)
* ci: Abort previous concurrent tests
* ci: Update de base docker image on build
* fix(tests): Fix typo in setup-machine script
* fix(tests): Amend upgrade tests (#146)
* fix: Update services to wait for network (#151)
* fix: Update uraft service to restart on failure (#153)
* fix: Release compilation on GCC 14 (#148)
* feat(chunkserver): Extend IDiskManager interface (#143)
* test: Verify master loading legacy metadata
* feat: Load legacy metadata file
* fix: Allow legacy sessions download
* fix: Allow legacy metadata download

+++ tests/tools/saunafsXX.sh
@@ -8,7 +8,10 @@
 SAUNAFSXX_TAG="4.1.0"
 
 install_saunafsXX() {
-	rm -rf "${SAUNAFSXX_DIR}"
+	if [ "${SAUNAFSXX_DIR:?}" == "/" ]; then
+		test_fail "Cowardly refusing to perform potentially destructive operation on /"
+	fi
+	rm -rf "${SAUNAFSXX_DIR:?}"
 	mkdir -p "${SAUNAFSXX_DIR}"
 	local distro="$(lsb_release -si)"
 	case "${distro}" in
@@ -17,20 +20,30 @@ install_saunafsXX() {
 		local codename="$(lsb_release -sc | tail -1)"
 		local release="$(lsb_release -sr | tail -1)"
 		mkdir -p "${TEMP_DIR}/apt/apt.conf.d"
-		mkdir -p "${TEMP_DIR}/apt/var/lib/apt/partial"
+		mkdir -p "${TEMP_DIR}/apt/auth.conf.d"
+		mkdir -p "${TEMP_DIR}/apt/sources.list.d"
 		mkdir -p "${TEMP_DIR}/apt/var/cache/apt/archives/partial"
+		mkdir -p "${TEMP_DIR}/apt/var/lib/apt/partial"
 		mkdir -p "${TEMP_DIR}/apt/var/lib/dpkg"
 		mkdir -p "${TEMP_DIR}/usr/share/keyrings"
 		cp /var/lib/dpkg/status "${TEMP_DIR}/apt/var/lib/dpkg/status"
+		if [ -f /tmp/saunafs-test-auth.conf ]; then
+			cp /tmp/saunafs-test-auth.conf "${TEMP_DIR}/apt/auth.conf.d/saunafs.conf"
+		fi
 		cat >"${TEMP_DIR}/apt/apt.conf" <<END
-Dir::State "${TEMP_DIR}/apt/var/lib/apt";
-Dir::State::status "${TEMP_DIR}/apt/var/lib/dpkg/status";
-Dir::Etc::SourceList "${TEMP_DIR}/apt/saunafs.list";
 Dir::Cache "${TEMP_DIR}/apt/var/cache/apt";
 Dir::Etc::Parts "${TEMP_DIR}/apt/apt.conf.d";
+Dir::Etc::netrc "${TEMP_DIR}/apt/auth.conf";
+Dir::Etc::netrcparts "${TEMP_DIR}/apt/auth.conf.d";
+Dir::Etc::SourceList "${TEMP_DIR}/apt/saunafs.list";
+Dir::Etc::SourceParts "${TEMP_DIR}/apt/sources.list.d";
+Dir::State "${TEMP_DIR}/apt/var/lib/apt";
+Dir::State::status "${TEMP_DIR}/apt/var/lib/dpkg/status";
 END
+		curl -sSL "https://keyserver.ubuntu.com/pks/lookup?op=get&search=0xa80b96e2c79457d4" | \
+			gpg --dearmor -o "${TEMP_DIR}/apt/saunafs-archive-keyring.gpg"
 		local destdir="${TEMP_DIR}/apt/var/cache/apt/archives"
-		echo "deb [arch=amd64 signed-by=/usr/share/keyrings/saunafs-archive-keyring.gpg] https://repo.saunafs.com/repository/saunafs-${distro_id}-${release}/ ${codename} main" >"${TEMP_DIR}/apt/saunafs.list"
+		echo "deb [arch=amd64 signed-by=${TEMP_DIR}/apt/saunafs-archive-keyring.gpg] https://repo.saunafs.com/repository/saunafs-${distro_id}-${release}/ ${codename} main" >"${TEMP_DIR}/apt/saunafs.list"
 		env APT_CONFIG="${TEMP_DIR}/apt/apt.conf" apt-get update
 		env APT_CONFIG="${TEMP_DIR}/apt/apt.conf" apt-get install --yes libyaml-cpp*
 		SAUNAFSXX_TAG_APT="4.1.0-20240509-152518-stable-main-a7cb5669"
@@ -115,7 +128,7 @@ test_saunafsXX_executables() {
 
 saunafsXX_chunkserver_daemon() {
 	# shellcheck disable=SC2154
-	"${SAUNAFSXX_DIR}/sbin/sfschunkserver" -c "${saunafs_info_[chunkserver${1_cfg}]}" "${2}" | cat
+	"${SAUNAFSXX_DIR}/sbin/sfschunkserver" -c "${saunafs_info_["chunkserver${1}_cfg"]}" "${2}" | cat
 	return "${PIPESTATUS[0]}"
 }
